Refactor semantic manifest handling: introduce `dbt_semantic.py` for modularity and reuse, streamline path resolution and grain processing.
